% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runnonmem.R
\name{runnonmem}
\alias{runnonmem}
\title{Run NONMEM}
\usage{
runnonmem(
  name,
  dir = getwd(),
  file = NULL,
  nmfe_location = getOption("runnonmem_nmfe_location")
)
}
\arguments{
\item{name}{name of the model to run located in \code{dir}. Accepted extension are: ".mod", ".ctl" and ".txt"}

\item{dir}{project directory where the file in \code{name} is and where \code{NONMEM} will be executed}

\item{file}{full name file with directory, alternative to \code{name} and \code{dir} arguments.}

\item{nmfe_location}{location of the \code{nmfeXX.bat} script, usually available in the NONMEM installation directory}
}
\value{
call for side effects. Open a shell where NONMEM runs
}
\description{
Run NONMEM from R. Creates a \code{.bat} file in a temporary directory and executes it from R.
}
\examples{
\dontrun{
# First, define where is NONMEM is installed
options(runnonmem_nmfe_location = "C:/nmXXXXX/util/nmfeXX.bat")

# Then launch a run
# If the control stream is in the working directory, just execute
runnonmem("mod001.mod")

# Or specify the project directory apart
runnonmem("mod001.mod", "/my/working/directory")

# Or full path to model object
runnonmem(file = "C:/my/working/directory.mod001.mod")
}


}
